Refactor unit tests & code structure
